title added

+++ Employee/views/v1/views.py
@@ -5129,7 +5129,8 @@ def update_workingschedule(request):
                 employee_id=employee,
                 is_weekend=True
             )
-            working_schedule_to_del.delete()
+            if working_schedule_to_del:
+                working_schedule_to_del.delete()
             working_schedule = EmployeDailySchedule.objects.create(
                 business_id=business_id,
                 employee_id=employee,
@@ -5138,8 +5139,9 @@ def update_workingschedule(request):
                 is_vacation=False
             )
         working_schedule = EmployeDailySchedule.objects.filter(
-            employee__id__in=week_end_employee,
-            is_weekend=True
+            # employee__id__in=week_end_employee,
+            is_weekend=True,
+            date=date
         )
         serializers = ScheduleSerializer(working_schedule, context={'request': request}, many=True)
         return Response(
